+++ revenuecat_examples/purchase_tester/lib/src/app.dart
@@ -1,12 +1,12 @@
 import 'dart:async';
-import 'dart:developer';
 
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:purchases_flutter/purchases_flutter.dart';
-import 'package:purchases_ui_flutter/purchases_ui_flutter.dart';
 
-import './constant.dart';
+import 'constant.dart';
+import 'cats.dart';
+import 'upsell.dart';
 
 class PurchaseTester extends StatelessWidget {
   const PurchaseTester({Key? key}) : super(key: key);
@@ -86,256 +86,4 @@ class _MyAppState extends State<InitialScreen> {
       }
     }
   }
-}
-
-// ignore: public_member_api_docs
-class UpsellScreen extends StatefulWidget {
-  const UpsellScreen({Key? key}) : super(key: key);
-
-  @override
-  State<StatefulWidget> createState() => _UpsellScreenState();
-}
-
-class _UpsellScreenState extends State<UpsellScreen> {
-  Offerings? _offerings;
-
-  @override
-  void initState() {
-    super.initState();
-    fetchData();
-  }
-
-  Future<void> fetchData() async {
-    Offerings? offerings;
-    try {
-      offerings = await Purchases.getOfferings();
-    } on PlatformException catch (e) {
-      print(e);
-    }
-
-    if (!mounted) return;
-
-    setState(() {
-      _offerings = offerings;
-    });
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    if (_offerings != null) {
-      final offering = _offerings!.current;
-      if (offering != null) {
-        List<Widget> buttonThings = offering.availablePackages
-            .map((package) {
-              List<Widget> buttons = [
-                _PurchaseButton(package: package),
-                _PurchaseStoreProductButton(storeProduct: package.storeProduct)
-              ];
-
-              List<Widget> optionButtons =
-                  (package.storeProduct.subscriptionOptions?.map((e) {
-                            return _PurchaseSubscriptionOptionButton(option: e);
-                          }) ??
-                          [])
-                      .toList();
-
-              buttons.addAll(optionButtons);
-
-              return buttons;
-            })
-            .expand((i) => i)
-            .toList();
-
-        buttonThings.add(ElevatedButton(
-          onPressed: () async {
-            final paywallResult = await RevenueCatUI.presentPaywall();
-            log('Paywall result: $paywallResult');
-          },
-          child: const Text('Present paywall'),
-        ));
-
-        return Scaffold(
-          appBar: AppBar(title: const Text('Upsell Screen')),
-          body: Center(
-            child: Column(
-              mainAxisSize: MainAxisSize.min,
-              children: buttonThings,
-            ),
-          ),
-        );
-      }
-    }
-    return Scaffold(
-      appBar: AppBar(title: const Text('Upsell Screen')),
-      body: const Center(
-        child: Text('Loading...'),
-      ),
-    );
-  }
-}
-
-class _PurchaseButton extends StatelessWidget {
-  final Package package;
-
-  // ignore: public_member_api_docs
-  const _PurchaseButton({Key? key, required this.package}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) => ElevatedButton(
-        onPressed: () async {
-          try {
-            final customerInfo = await Purchases.purchasePackage(package);
-            final isPro =
-                customerInfo.entitlements.active.containsKey(entitlementKey);
-            if (isPro) {
-              Navigator.pushReplacement(
-                context,
-                MaterialPageRoute(builder: (context) => const CatsScreen()),
-              );
-            }
-          } on PlatformException catch (e) {
-            final errorCode = PurchasesErrorHelper.getErrorCode(e);
-            if (errorCode == PurchasesErrorCode.purchaseCancelledError) {
-              print('User cancelled');
-            } else if (errorCode ==
-                PurchasesErrorCode.purchaseNotAllowedError) {
-              print('User not allowed to purchase');
-            } else if (errorCode == PurchasesErrorCode.paymentPendingError) {
-              print('Payment is pending');
-            }
-          }
-          Navigator.pushReplacement(
-            context,
-            MaterialPageRoute(builder: (context) => const InitialScreen()),
-          );
-        },
-        child: Text(
-            'Buy Package: ${package.storeProduct.subscriptionPeriod ?? package.storeProduct.title}\n${package.storeProduct.priceString}'),
-      );
-}
-
-class _PurchaseStoreProductButton extends StatelessWidget {
-  final StoreProduct storeProduct;
-
-  // ignore: public_member_api_docs
-  const _PurchaseStoreProductButton({Key? key, required this.storeProduct})
-      : super(key: key);
-
-  @override
-  Widget build(BuildContext context) => ElevatedButton(
-        onPressed: () async {
-          try {
-            final customerInfo =
-                await Purchases.purchaseStoreProduct(storeProduct);
-            final isPro =
-                customerInfo.entitlements.active.containsKey(entitlementKey);
-            if (isPro) {
-             Navigator.pushReplacement(
-                context,
-                MaterialPageRoute(builder: (context) => const CatsScreen()),
-              );
-            }
-          } on PlatformException catch (e) {
-            final errorCode = PurchasesErrorHelper.getErrorCode(e);
-            if (errorCode == PurchasesErrorCode.purchaseCancelledError) {
-              print('User cancelled');
-            } else if (errorCode ==
-                PurchasesErrorCode.purchaseNotAllowedError) {
-              print('User not allowed to purchase');
-            } else if (errorCode == PurchasesErrorCode.paymentPendingError) {
-              print('Payment is pending');
-            }
-          }
-          Navigator.pushReplacement(
-            context,
-            MaterialPageRoute(builder: (context) => const InitialScreen()),
-          );
-        },
-        child: Text(
-            'Buy StoreProduct (${storeProduct.productCategory}): ${storeProduct.subscriptionPeriod ?? storeProduct.title}\n${storeProduct.priceString}'),
-      );
-}
-
-class _PurchaseSubscriptionOptionButton extends StatelessWidget {
-  final SubscriptionOption option;
-
-  // ignore: public_member_api_docs
-  const _PurchaseSubscriptionOptionButton({Key? key, required this.option})
-      : super(key: key);
-
-  @override
-  Widget build(BuildContext context) => ElevatedButton(
-        onPressed: () async {
-          try {
-            final customerInfo =
-                await Purchases.purchaseSubscriptionOption(option);
-            final isPro =
-                customerInfo.entitlements.active.containsKey(entitlementKey);
-            if (isPro) {
-              Navigator.pushReplacement(
-                context,
-                MaterialPageRoute(builder: (context) => const CatsScreen()),
-              );
-            }
-          } on PlatformException catch (e) {
-            final errorCode = PurchasesErrorHelper.getErrorCode(e);
-            if (errorCode == PurchasesErrorCode.purchaseCancelledError) {
-              print('User cancelled');
-            } else if (errorCode ==
-                PurchasesErrorCode.purchaseNotAllowedError) {
-              print('User not allowed to purchase');
-            } else if (errorCode == PurchasesErrorCode.paymentPendingError) {
-              print('Payment is pending');
-            }
-          }
-        },
-        child:
-            Text('Buy Option: - (${option.id}\n${option.pricingPhases.map((e) {
-          return '${e.price.formatted} for ${e.billingPeriod?.value} ${e.billingPeriod?.unit}';
-        }).join(' -> ')})'),
-      );
-}
-
-// ignore: public_member_api_docs
-class CatsScreen extends StatelessWidget {
-  const CatsScreen({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) => Scaffold(
-        appBar: AppBar(title: const Text('Cats Screen')),
-        body: Center(
-            child: Column(
-          mainAxisSize: MainAxisSize.min,
-          children: [
-            const Text('User is pro'),
-            ElevatedButton(
-              onPressed: () async {
-                try {
-                  final customerInfo = await Purchases.getCustomerInfo();
-                  final refundStatus =
-                      await Purchases.beginRefundRequestForEntitlement(
-                          customerInfo.entitlements.active[entitlementKey]!);
-                  print('Refund request successful with status: $refundStatus');
-                } catch (e) {
-                  print('Refund request exception: $e');
-                }
-                Navigator.pushReplacement(
-                  context,
-                  MaterialPageRoute(
-                      builder: (context) => const InitialScreen()),
-                );
-              },
-              child: const Text('Begin refund for pro entitlement'),
-            ),
-            ElevatedButton(
-              onPressed: () async {
-                // In order to defer in-app messages so they're only shown when this button is pressed, you must configure
-                // the SDK with `configuration.shouldShowInAppMessagesAutomatically = false;`
-                Purchases.showInAppMessages(types: {InAppMessageType.billingIssue, InAppMessageType.priceIncreaseConsent, InAppMessageType.generic});
-              },
-              child: const Text('Show In-App Messages'),
-            ),
-          ],
-        )),
-      );
 }
